FileBackend should close all the files it opens
This was causing a number of apparently unrelated exceptions with
our website [1] which uses vanilla-rb [2] when the OS ran out of
file handles.

A simple way to reproduce the problem is to reduce the number of
file handles available in your OS, e.g. using `ulimit -n 16` on OSX.

I think there are at least a couple of possible failure modes, but
the one we've definitely seen is where the call to `Dir::[]` in
`FileBackend#snip_paths` erroneously returns an empty array because
the OS has run out of file handles. It seems odd that this just
silently fails and doesn't raise an exception. I might submit a
bug report to Ruby about this. You can see a simple demonstration
of the problem here [3].

I can imagine scenarios in which the call to `File.new` in
`FileBackend#load_snip_from_path` might raise the following exception:

    Errno::EMFILE: Too many open files

We haven't actually seen this exception occur, but it's possible that's
because vanilla-rb is swallowing it.

I haven't bothered to write an explicit test for this change, because
it feels as if it's just good practice in Ruby to ensure all files
are closed.

It doesn't look as if any of the other backend types suffer from the
same problem, so I haven't touched them.

Many thanks to @chrisroos for help in diagnosing this problem.

Fixes lazyatom/vanilla-rb#10.

[1]: https://github.com/freerange/site
[2]: https://github.com/lazyatom/vanilla-rb
[3]: https://gist.github.com/floehopper/3ee80293e8fee340f0db816977a9b2e9